refactor(UserInfoRepository): Modify PK type
Long -> Integer

+++ src/main/java/mathrone/backend/repository/UserInfoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
+public interface UserInfoRepository extends JpaRepository<UserInfo, Integer> {
 
     UserInfo findByUserId(Integer userId);
